code snippets/requirements.txt
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-google-genai
faiss-cpu
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
notebook
ipykernel